requirements.txt
# GPU-compatible dependencies for CERN Jet Anomaly Detection
# Install with: pip install -r requirements.txt
# For CUDA 11.8: pip install torch>=2.0.0 torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch>=2.0.0 torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/cu121

torch>=2.0.0
torchvision>=0.15.0

numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.65.0
seaborn>=0.12.0

# PCA is used as the default for dimensionality reduction.
# umap-learn is optional and was removed due to build issues on some systems.
# If you wish to use UMAP, you must manually install llvmlite, numba, and umap-learn.


pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
